# css: Improve type-checking performance of `CSSVarFunction`