--- a/.changeset/nervous-cougars-glow.md
+++ b/.changeset/nervous-cougars-glow.md
@@ -1,0 +1,5 @@
+---
+'@vanilla-extract/css': patch
+---
+
+`css`: Improve type-checking performance of string literal types that include CSS variables