--- a/.changeset/two-kids-relax.md
+++ b/.changeset/two-kids-relax.md
@@ -1,0 +1,5 @@
+---
+'@vanilla-extract/private': patch
+---
+
+Simplify `CSSVarFunction` type to improve type-checking performance