--- a/benchmarks/README.md
+++ b/benchmarks/README.md
@@ -1,0 +1,11 @@
+# Benchmarks
+
+This package contains scripts for benchmarking the runtime and type-checking performance of Vanilla Extract APIs.
+
+## Running a Benchmark
+
+Run a benchmark by executing it with `tsx`:
+
+```sh
+pnpm tsx ./src/css-var-string-literal.ts
+```