--- a/pnpm-workspace.yaml
+++ b/pnpm-workspace.yaml
@@ -7,3 +7,4 @@ packages:
   - 'site'
   - 'examples/*'
   - 'tests'
+  - 'benchmarks'