@jojojojojoj5564656465465 2025-03-28
```ts
const nbCards = createVar({
  syntax: '<integer>',
  initialValue: '10',
  inherits: false,
})
const jfk = calc.divide('60rem', nbCards)
```

Why it's not working ?