@askoufis 2025-03-28
> ```ts
> const nbCards = createVar({
>   syntax: '<integer>',
>   initialValue: '10',
>   inherits: false,
> })
> const jfk = calc.divide('60rem', nbCards)
> ```
> 
> Why it's not working ?

Please don't revive old PRs unless they're related to your issue. I'd prefer you make a separate issue with a minimal reproduction.